make the index and the status request works correctly

+++ src/Models/Todo/TodoDto.cs
@@ -7,7 +7,7 @@ public class TodoDto {
     public string? Title { get; set; }
     public string Context { get; set; }
     public string username { get; set; }
-    public int index { get; set; }
+    public float index { get; set; }
     public int ProjectId { get; set; }
     public string Status { get; set; } = "TODO";
 }
